refactor archi and add code  documentation
