+++ Src/produit.c
@@ -1,10 +1,27 @@
+/** 
+ * @file produit.c
+ *
+ * @brief fonctions de gestion des produits pour un système de gestion de stock.
+ * Les opérations incluent l'ajout, l'affichage, la recherche, la modification et la suppression de produits.
+ *
+ * @author Oussama
+ * @date 2024-06-09
+ * @version 1.0
+ */
 // produit.c
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include "produit.h"
 
-// Ajouter un nouveau produit
+/**
+ * @brief Ajoute un nouveau produit au fichier "stock.dat".
+ *
+ * Cette fonction demande à l'utilisateur de saisir les informations d'un produit
+ * (nom, identifiant, quantité, prix), puis enregistre ces informations dans le fichier
+ * binaire "stock.dat". Si le fichier ne peut pas être ouvert, un message d'erreur est affiché.
+ *
+ */
 void ajouterProduit() {
     Produit p;
     FILE *f = fopen("stock.dat", "ab");
@@ -30,7 +47,16 @@ void ajouterProduit() {
     printf(" Produit ajouté !\n");
 }
 
-// Affiche tous les produits du fichier
+/**
+ * @brief Affiche la liste des produits stockés dans le fichier "stock.dat".
+ *
+ * Cette fonction ouvre le fichier binaire "stock.dat" en mode lecture,
+ * lit chaque structure Produit enregistrée, et affiche ses informations
+ * (ID, nom, quantité, prix) sur la sortie standard. Si le fichier n'existe pas
+ * ou ne peut pas être ouvert, un message d'erreur est affiché.
+ *
+ */
+
 void afficherProduits() {
     Produit p;
     FILE *f = fopen("stock.dat", "rb");
@@ -46,7 +72,15 @@ void afficherProduits() {
     fclose(f);
 }
 
-// Recherche un produit par nom
+/**
+ * @brief Recherche un produit dans le fichier "stock.dat" par son nom et affiche ses informations si trouvé.
+ *
+ * Cette fonction demande à l'utilisateur de saisir le nom d'un produit, puis parcourt le fichier binaire "stock.dat"
+ * pour rechercher un produit correspondant. Si le produit est trouvé, ses informations (ID, quantité, prix) sont affichées.
+ * Si aucun produit n'est trouvé, un message approprié est affiché.
+ *
+ */
+
 void chercherProduit() {
     Produit p;
     char nomRecherche[50];
@@ -73,7 +107,22 @@ void chercherProduit() {
     fclose(f);
 }
 
-// Modifie un produit existant (par ID)
+/**
+ * @brief Modifie les informations d'un produit existant dans le fichier "stock.dat".
+ *
+ * Cette fonction recherche un produit par son identifiant (ID) dans le fichier binaire "stock.dat".
+ * Si le produit est trouvé, elle permet à l'utilisateur de modifier le nom, la quantité et le prix du produit.
+ * Les modifications sont enregistrées directement dans le fichier.
+ *
+ * Étapes principales :
+ * 1. Demande à l'utilisateur l'ID du produit à modifier.
+ * 2. Ouvre le fichier "stock.dat" en mode lecture/écriture binaire.
+ * 3. Parcourt le fichier à la recherche du produit correspondant à l'ID.
+ * 4. Si trouvé, demande les nouvelles informations et met à jour le produit dans le fichier.
+ * 5. Affiche un message de confirmation ou d'erreur selon le résultat.
+ *
+ */
+
 void modifierProduit() {
     Produit p;
     int idRecherche;
@@ -113,7 +162,19 @@ void modifierProduit() {
     fclose(f);
 }
 
-// Supprime un produit (recopie sans le produit ciblé)
+/**
+ * @brief Supprime un produit du fichier "stock.dat" en fonction de son identifiant.
+ *
+ * Cette fonction demande à l'utilisateur de saisir l'identifiant du produit à supprimer.
+ * Elle lit ensuite tous les produits du fichier binaire "stock.dat" et copie ceux dont
+ * l'identifiant est différent dans un fichier temporaire "temp.dat". Si le produit à
+ * supprimer est trouvé, il n'est pas copié dans le fichier temporaire. À la fin de
+ * l'opération, le fichier original est remplacé par le fichier temporaire.
+ *
+ * Affiche un message indiquant si le produit a été supprimé ou non trouvé.
+ *
+ */
+
 void supprimerProduit() {
     Produit p;
     int idSupp;
